A06の変更点の更新

+++ JC21PS/src/jp/co/jcps/A06/JoinApprovalSaveServlet.java
@@ -43,7 +43,7 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 		// リクエストから情報を取得する
 		boolean approvalFlg = request.getParameter("approvalFlg").equals("true");
 		// TODO: リクエストから承認・否認するユーザーのユーザーIDを取得しなさい
-		String registUserId = ;
+		String registUserId =(String)request.getSession() .getAttribute("registUserID");
 
 		// セッションからログイン中のユーザーの部長クラブIDを取得する
 		String leaderClubId = (String) request.getSession().getAttribute("leaderClubId");
@@ -52,11 +52,11 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 			if (approvalFlg) {
 				// 承認する場合
 				// TODO: 部員登録申請を承認する場合のみ実行する処理（メソッド）を呼び出しなさい。
-
+				 saveClubMember(leaderClubId,registUserId);
 
 			}
-			//TODO: 部員登録申請を承認する場合および拒否する場合どちらも実行する処理（メソッド）を呼び出しなさい。
-
+				//TODO: 部員登録申請を承認する場合および拒否する場合どちらも実行する処理（メソッド）を呼び出しなさい。
+			 	deleteJoinRequest(leaderClubId,registUserId);
 		}catch(Exception e) {
 			request.getRequestDispatcher("ERROR/Error.jsp").forward(request, response);
 		}
@@ -74,12 +74,13 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 
 		//SQLを宣言
 		// TODO: SQL文を完成させなさい。
-		String sql = "INSERT INTO ";
+		String sql = "INSERT INTO turn_club_menber VALUES(registClubID,registUserID)";
 
 		// SQLに埋め込むパラメータリストを定義
 		List<String> paramList = new ArrayList<String>();
 		// TODO: SQLに埋め込む値をparamListに設定しなさい。
-
+		paramList.add(registClubId);
+		paramList.add(registUserId);
 
 
 		// SQLを実行し結果を取得
@@ -97,12 +98,13 @@ public class JoinApprovalSaveServlet extends HttpServlet {
 
 		//SQLを宣言
 		// TODO: SQL文を完成させなさい。
-		String sql = "DELETE FROM ";
+		String sql = "DELETE FROM trn_join_request WHERE  trn_join_request.user_id = registUserID AND trn_join_request.club_ID = registClubID ";
 
 		// SQLに埋め込むパラメータリストを定義
 		List<String> paramList = new ArrayList<String>();
 		// TODO: SQLに埋め込む値をparamListに設定しなさい。
-
+		paramList.add(registClubId);
+		paramList.add(registUserId);
 
 		// SQLを実行し結果を取得
 		DBConnection db = new DBConnection();
